backend/pyproject.toml
[project]
name = "lyricguess-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "psycopg2-binary",
    "ytmusicapi",
    "python-dotenv",
]
